object restructure nearly done. does not build presently.
